refactor: Break up wsm.go into per command files and move main.go to project root
