+++ cmd/deploy/deploy.go
@@ -1,0 +1,1 @@
+package deploy
