+++ main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/wandb/wsm/cmd"
+
+func main() {
+	cmd.Execute()
+}
